Config WIDGET_GARDEN_URL to 0.0.0.0:3001 in development mode
Was getting annoying to keep changing the constant value for this
in the widget model, so I changed it to default to localhost
instead of Heroku when you run in dev mode. Anybody can customize
this for themselves by tweaking their development.rb, or if folks
are all using different values here we could put it in a new file
that is not under version control but gets loaded during bootup,
similar to config/developers/logins.rb in core.