Add DropTarget model to Rails & refactor seeds
The DropTarget model acts as an intermediary to WebTemplates and
Widgets. A WebTemplate has many drop_targets and a DropTarget has many
widgets.

This change also prompted me to rewrite the way that defaults work. Now
instead of having Ruby classes for different default pages it is all the
defaults are now stored as YAML. It is really only used in the seeding
process at the moment, which I decided to brute force instead of
cluttering the models with accepts_nested_attributes_for.

This change should allow for an easier time moving widgets to different
drop targets in the Ember app.