Adds example web template request spec
Hey, @crissmancd!

I spent a little time tonight seeing if i could make it easier for you to
test things, and I think I found a good way! I am hoping this will mean
that more people (you) are writing specs and that it is easy :)

I wrote an example spec of the kind that I think will be useful for you.
It is similar to the integration specs we wrote together for
g5-phone-number-service.

It should be very easy for you to define what layout/theme/widgets a
page uses, configure the page, and render the page without having to
click around a lot in the browser.

The coolest part is that you can generate the pages and take
screenshots! I included instructions in the spec file. You can also view
them as HTML using a debugger. We should pair on it!

This is to help solve big problems. You can verify that content and
selectors are present in spec files. "Everything looks awesome"
verifications are aided by screenshots and automation of building. This
is not a full replacement for clicking around in the browser like a real
user would.