Create or update clients and locations
I did not try to run any of this code. This commit is to illustrate what
I was thinking about doing. I just sketched it out real quick.
@jessicard please make sure it works :P Also we should talk about why I
used this approach if it is not clear. I also tried to name somethings
more clearly. FYI uf2 is shorthand for microformats2.